New: Allow documents to be viewed in presentation mode
