+++ src/lib/viewers/doc/DocLoader.js
@@ -5,9 +5,81 @@ import PresentationViewer from './PresentationViewer';
 import RepStatus from '../../RepStatus';
 import { ORIGINAL_REP_NAME, STATUS_SUCCESS } from '../../constants';
 
-// Order of the viewers matters. Prefer original before others. Go from specific to general.
-// For example, a pdf file can be previewed both natively (majority use case) using the original
-// representation but can fallback to using the pdf representation (for watermarked versions).
+const DOCUMENT_FORMATS = [
+    'as',
+    'as3',
+    'asm',
+    'bat',
+    'c',
+    'cc',
+    'cmake',
+    'cpp',
+    'cs',
+    'css',
+    'csv',
+    'cxx',
+    'diff',
+    'doc',
+    'docx',
+    'erb',
+    'gdoc',
+    'groovy',
+    'gsheet',
+    'h',
+    'haml',
+    'hh',
+    'htm',
+    'html',
+    'java',
+    'js',
+    'less',
+    'log',
+    'm',
+    'make',
+    'md',
+    'ml',
+    'mm',
+    'msg',
+    'odp',
+    'ods',
+    'odt',
+    'pdf',
+    'php',
+    'pl',
+    'plist',
+    'ppt',
+    'pptx',
+    'properties',
+    'py',
+    'rb',
+    'rst',
+    'rtf',
+    'sass',
+    'scala',
+    'scm',
+    'script',
+    'sh',
+    'sml',
+    'sql',
+    'tsv',
+    'txt',
+    'vi',
+    'vim',
+    'webdoc',
+    'wpd',
+    'xhtml',
+    'xls',
+    'xlsm',
+    'xlsx',
+    'xml',
+    'xsd',
+    'xsl',
+    'yaml'
+];
+
+// Order of the viewers matters. For example, a PDF file can be previewed by using the preferred optimized 'pdf' rep
+// or the original as a fallback. Additionally, we include multiple entries for the presentation viewer so that it can be
+// used by other docoument types if the document viewer is disabled.
 const VIEWERS = [
     {
         NAME: 'Presentation',
@@ -19,81 +91,26 @@ const VIEWERS = [
         NAME: 'Document',
         CONSTRUCTOR: DocumentViewer,
         REP: 'pdf',
-        EXT: [
-            'as',
-            'as3',
-            'asm',
-            'bat',
-            'c',
-            'cc',
-            'cmake',
-            'cpp',
-            'cs',
-            'css',
-            'csv',
-            'cxx',
-            'diff',
-            'doc',
-            'docx',
-            'erb',
-            'gdoc',
-            'groovy',
-            'gsheet',
-            'h',
-            'haml',
-            'hh',
-            'htm',
-            'html',
-            'java',
-            'js',
-            'less',
-            'log',
-            'm',
-            'make',
-            'md',
-            'ml',
-            'mm',
-            'msg',
-            'odp',
-            'ods',
-            'odt',
-            'pdf',
-            'php',
-            'pl',
-            'plist',
-            'ppt',
-            'pptx',
-            'properties',
-            'py',
-            'rb',
-            'rst',
-            'rtf',
-            'sass',
-            'scala',
-            'scm',
-            'script',
-            'sh',
-            'sml',
-            'sql',
-            'tsv',
-            'txt',
-            'vi',
-            'vim',
-            'webdoc',
-            'wpd',
-            'xhtml',
-            'xls',
-            'xlsm',
-            'xlsx',
-            'xml',
-            'xsd',
-            'xsl',
-            'yaml'
-        ]
+        EXT: DOCUMENT_FORMATS
+    },
+    // Allows other document types to use the presentation viewer when the document viewer is disabled.
+    {
+        NAME: 'Presentation',
+        CONSTRUCTOR: PresentationViewer,
+        REP: 'pdf',
+        EXT: DOCUMENT_FORMATS
     },
     {
         NAME: 'Document',
         CONSTRUCTOR: DocumentViewer,
+        REP: ORIGINAL_REP_NAME,
+        EXT: ['pdf', 'lcdpdf']
+    },
+    //  Allows PDFs and lcpdf files that only have an original rep
+    // to use the presentation viewer when the document viewer is disabled.
+    {
+        NAME: 'Presentation',
+        CONSTRUCTOR: PresentationViewer,
         REP: ORIGINAL_REP_NAME,
         EXT: ['pdf', 'lcdpdf']
     }
