+++ src/lib/viewers/doc/DocumentViewer.js
@@ -2,29 +2,8 @@ import autobind from 'autobind-decorator';
 import DocBaseViewer from './DocBaseViewer';
 import DocPreloader from './DocPreloader';
 import fullscreen from '../../Fullscreen';
-import {
-    ICON_FILE_DOCUMENT,
-    ICON_FILE_PDF,
-    ICON_FILE_SPREADSHEET,
-    ICON_FILE_WORD,
-    ICON_FULLSCREEN_IN,
-    ICON_FULLSCREEN_OUT,
-    ICON_ZOOM_IN,
-    ICON_ZOOM_OUT
-} from '../../icons/icons';
+import { ICON_FULLSCREEN_IN, ICON_FULLSCREEN_OUT, ICON_ZOOM_IN, ICON_ZOOM_OUT } from '../../icons/icons';
 import './Document.scss';
-
-const LOADING_ICON_MAP = {
-    csv: ICON_FILE_SPREADSHEET,
-    doc: ICON_FILE_WORD,
-    docx: ICON_FILE_WORD,
-    gdoc: ICON_FILE_WORD,
-    gsheet: ICON_FILE_SPREADSHEET,
-    pdf: ICON_FILE_PDF,
-    xls: ICON_FILE_SPREADSHEET,
-    xlsm: ICON_FILE_SPREADSHEET,
-    xlsx: ICON_FILE_SPREADSHEET
-};
 
 @autobind
 class DocumentViewer extends DocBaseViewer {
@@ -36,9 +15,6 @@ class DocumentViewer extends DocBaseViewer {
      * @inheritdoc
      */
     setup() {
-        const fileExt = this.options.file.extension;
-        this.fileLoadingIcon = LOADING_ICON_MAP[fileExt] || ICON_FILE_DOCUMENT;
-
         // Call super() to set up common layout
         super.setup();
         this.docEl.classList.add('bp-doc-document');
